pyproject.toml
[project]
name = "md-img-uri"
version = "0.1.0"
description = "Embed images into Markdown as data URIs"
readme = "README.md"
authors = [
    { name = "Ralph Schmieder", email = "rschmied@cisco.com" }
]
requires-python = ">=3.12"
dependencies = ["pillow>=10.0.0"]
license = { text = "MIT" }
keywords = [
    "markdown",
    "image",
    "data-uri",
    "base64",
    "svg",
    "embedding",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]

[project.urls]
Repository = "https://github.com/rschmied/md-img-uri"
Issues = "https://github.com/rschmied/md-img-uri/issues"

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[project.scripts]
md-img-uri = "md_img_uri.cli:main"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
